refactor(ricer::locate): simplify module structure
