+++ src/locate.rs
@@ -23,18 +23,18 @@
 //!
 //! [xdg]: https://specifications.freedesktop.org/basedir-spec/latest/
 
-mod error;
-mod layout;
-
-#[doc(inline)]
-pub use error::*;
-pub use layout::*;
-
+use directories::ProjectDirs;
 use log::{debug, trace};
 use std::path::{Path, PathBuf};
 
 #[cfg(test)]
 use mockall::automock;
+
+#[derive(Debug, PartialEq, Eq, thiserror::Error)]
+pub enum LocateError {
+    #[error("Cannot determine path to home directory")]
+    NoWayHome,
+}
 
 /// Configuration data locator.
 #[cfg_attr(test, automock)]
@@ -106,5 +106,45 @@ impl Locator for DefaultLocator {
 
     fn repos_config(&self) -> &Path {
         self.repos_config.as_path()
+    }
+}
+
+/// Specify expected configuration directory layout.
+#[cfg_attr(test, automock)]
+pub trait DirLayout {
+    /// Absolute path to directory where configuration files will be stored.
+    fn config_dir(&self) -> &Path;
+
+    /// Absolute path to directory where repository data will be stored.
+    fn repo_dir(&self) -> &Path;
+}
+
+/// Configuration directory layout handler following [XDG Base Directory
+/// Specification][xdg].
+///
+/// # Invariants
+///
+/// 1. Caller must validate paths themselves.
+///
+/// [xdg]: https://specifications.freedesktop.org/basedir-spec/latest/
+pub struct XdgDirLayout {
+    layout: ProjectDirs,
+}
+
+impl XdgDirLayout {
+    pub fn layout() -> Result<Self, LocateError> {
+        trace!("Construct XDG Base Directory Specification layout handler");
+        let layout = ProjectDirs::from("com", "awkless", "ricer").ok_or(LocateError::NoWayHome)?;
+        Ok(Self { layout })
+    }
+}
+
+impl DirLayout for XdgDirLayout {
+    fn config_dir(&self) -> &Path {
+        self.layout.config_dir()
+    }
+
+    fn repo_dir(&self) -> &Path {
+        self.layout.data_dir()
     }
 }
